requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages

# Required for EmptyOperator and TimeDeltaSensorAsync used in the demo DAGs
apache-airflow-providers-standard
